requirements.txt
Django==5.2.7
djangorestframework==3.16.1
numpy==2.4.2
pandas==3.0.1
python-dotenv==1.2.2